README.md
InputFormMask
This is a Symfony application that upload the assets to a certain server and automate the process of filling large set of metadata in bulk. In the backend the admin users will create the necessary information associating the different features to a certain object or model, while in the frontend the users will upload the assets by choosing from the list of categories/products options the desired one, by doing this the asset will inherit all the necessary metadata, finally it will be generating a report of assets and metadata in XML format. A dashboard with all of the reports related to uploads, users, email notification and statistics is available.

This solution was built for automate the assets metadata process taking place in a famous technology firm.